Helper classes for the Wahlby-node moved to ..nodes.seg.wahlby-package.
